requirements.txt
datahub-client>=0.10.0
pandas>=1.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
fastavro>=1.7.0
pymongo>=4.0.0
boto3>=1.26.0
pytest>=7.0.0
black>=22.0.0
mypy>=1.0.0